source/ai-interview-bot/requirements.txt
streamlit
pypdf
torch
transformers<5
sentence-transformers
faiss-cpu
numpy
